Convert from wait time histograms to KDE in MCS

+++ include/file-reader/Incidents.hpp
@@ -26,6 +26,6 @@ class Incidents : public CSVReader {
         static Incidents instance;
         return instance;
     }
-    float timeDifferenceBetweenHeaders(const std::string& header1, const std::string& header2, const int index);
+    double timeDifferenceBetweenHeaders(const std::string& header1, const std::string& header2, const int index);
     std::vector<int> rowsWithinTimeFrame(const int month, const int day, const int windowSize);
 };
